Add loading states to API calls in previews
In all API calls within the Signature and Event Manager views, loading states have been added to provide better handling of states during the async calls. Prior to API call execution, the 'loading' state is set to true and upon completion, it is reset to false. Only when the 'loading' state is false, the relevant components are rendered.